Created a `linux` module and a `common` file for Linux and move the `deb_bundle` into it
